Refactoriza NavBar dividiendo en componentes más pequeños
